+++ plugins/modules/faas_deploy.py
@@ -47,6 +47,7 @@ options:
         required: false
         default: false
         type: bool
+        removed_in_version: '1.4.0'
     strategy:
         description: Whether to perform rolling update, or remove and re-create, one or more existing functions. This will deprecate the `replace` and `update` parameters in version 1.4.0.
         choices: [replace, update]
@@ -59,6 +60,7 @@ options:
         required: false
         default: true
         type: bool
+        removed_in_version: '1.4.0'
 
 requirements:
     - faas-cli >= 0.17.0
@@ -112,9 +114,9 @@ def main() -> None:
             'label': {'type': 'dict', 'required': False},
             'name': {'type': 'str', 'required': False},
             'regex': {'type': 'str', 'required': False},
-            'replace': {'type': 'bool', 'required': False},
+            'replace': {'type': 'bool', 'required': False, 'removed_in_version': '1.4.0'},
             'strategy': {'type': 'str', 'choices': ['replace', 'update'], 'default': 'update', 'required': False, 'new_in_version': '1.3.1'},
-            'update': {'type': 'bool', 'required': False, 'default': True},
+            'update': {'type': 'bool', 'required': False, 'default': True, 'removed_in_version': '1.4.0'},
         },
         mutually_exclusive=[('config_file', 'name'), ('strategy', 'replace'), ('strategy', 'update')],
         required_one_of=[('config_file', 'name')],
@@ -126,11 +128,6 @@ def main() -> None:
     if module.params.pop('strategy') == 'replace':
         flags.add('update')
         flags.add('replace')
-    else:
-        if module.params.get('update') is False:
-            flags.add('update')
-        if module.params.get('replace'):
-            flags.add('replace')
 
     # check args
     flags_args: tuple[set[str], dict] = universal.params_to_flags_args(module.params, module.argument_spec)